requirements.txt
discord.py
aiohttp
beautifulsoup4
requests
